+++ src/pages/activity/filters/Filters.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef } from "react";
 import { connect } from "react-redux";
-import { quakeFetch, updateSearchParams } from "../../actions";
-import useDarkMode from "../../customHooks/useDarkMode";
+import { quakeFetch, updateSearchParams } from "../../../redux/actions";
+import useDarkMode from "../../../utils/customHooks/useDarkMode";
 
 function Filters({
   quakeFetch,
