+++ src/pages/activity/results/Card.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { connect } from "react-redux";
-import { jumpViewport } from "../../../actions";
+import { jumpViewport } from "../../../redux/actions";
 import Arrow from "../../../images/Arrow_Active.png";
 var moment = require("moment");
