+++ src/pages/report/BugReport.js
@@ -3,7 +3,7 @@ import React from "react";
 const Report = () => {
   return (
     <div className="main-container no-scroll">
-      <p>WIP... Report</p>
+      <p>WIP... Bug Report</p>
     </div>
   );
 };
