+++ src/pages/resources/imageRow.js
