+++ src/pages/resources/listTypeSolid.js
