+++ src/pages/resources/titleContainer.js
