+++ src/redux/reducers/quakeReducer.js
